// src/products/dto/create-product-from-ai.dto.ts
import { ApiPropertyOptional } from '@nestjs/swagger';
import { IsArray, IsOptional, IsString } from 'class-validator';
import { CreateProductDto } from './create-product.dto';

export class CreateProductFromAiDto extends CreateProductDto {
  @ApiPropertyOptional({
    type: [String],
    description: 'List of image URLs generated by AI to be uploaded',
  })
  @IsOptional()
  @IsArray()
  @IsString({ each: true })
  imageUrls?: string[];

  @ApiPropertyOptional({
    description: 'AI generated description content',
  })
  @IsOptional()
  @IsString()
  aiDescription?: string;
}